提交

+++ Server/Src/ConfigData/ConfigStruct.h
@@ -151,7 +151,7 @@ struct StItemInfo
 	INT32  SellPrice;	//出售价格
 	UINT32 UseType;		//使用类型
 	UINT32 CarrerID;	//职业
-	INT32  StackMax;		//最大堆放数
+	INT64  StackMax;	//最大堆放数
 	INT32  Data1;		//参数1
 	INT32  Data2;		//参数2
 };
